+++ frontend/assets/LoginFormAsset.php
@@ -1,0 +1,19 @@
+<?php
+
+
+namespace frontend\assets;
+
+
+use yii\web\AssetBundle;
+use yii\web\JqueryAsset;
+
+class LoginFormAsset extends AssetBundle
+{
+    public $js = [
+        'js/loginForm.js'
+    ];
+    public $depends = [
+        JqueryAsset::class
+    ];
+
+}
