+++ frontend/views/landing/index.php
@@ -6,8 +6,9 @@
 use frontend\models\LoginForm;
 use yii\bootstrap\ActiveForm;
 use yii\bootstrap\Html;
-$this->registerJsFile('js/loginForm.js',
-    ['depends' => [\yii\web\JqueryAsset::className()]]);
+use frontend\assets\LoginFormAsset;
+
+LoginFormAsset::register($this);
 
 ?>
 <section class="enter-form form-modal" id="enter-form">
@@ -15,25 +16,24 @@ $this->registerJsFile('js/loginForm.js',
     <?php $form = ActiveForm::begin([
         'options' => [
             'tag' => false,
-            ],
+        ],
         'enableAjaxValidation' => true,
         'id' => 'loginForm',
         'validateOnBlur' => false,
         'validateOnChange' => false,
-//        'enableClientValidation'=>true,
     ]); ?>
-    <?= $form->field($loginForm, 'email',[
+    <?= $form->field($loginForm, 'email', [
         'options' => [
             'tag' => 'p',
-            ],
-        'labelOptions' => ['class' => 'form-modal-description'],
+        ],
+        'labelOptions' => ['class' => 'form-modal-description']
     ])->textInput([
         'autofocus' => true,
         'class' => 'enter-form-email input input-middle',
         'type' => 'email'
     ]) ?>
 
-    <?= $form->field($loginForm, 'password',[
+    <?= $form->field($loginForm, 'password', [
         'options' => [
             'tag' => 'p',
         ],
@@ -41,7 +41,77 @@ $this->registerJsFile('js/loginForm.js',
     ])->passwordInput(['class' => 'enter-form-email input input-middle']) ?>
     <?= Html::submitButton('Войти', ['class' => 'button']) ?>
     <?php ActiveForm::end(); ?>
-    <?= Html::button('Закрыть',['class' => 'form-modal-close', 'id' => 'close-modal'])?>
+    <?= Html::button('Закрыть', ['class' => 'form-modal-close', 'id' => 'close-modal']) ?>
 </section>
+<div class="landing-bottom">
+    <div class="landing-bottom-container">
+        <h2>Последние задания на сайте</h2>
+        <div class="landing-task">
+            <div class="landing-task-top task-courier"></div>
+            <div class="landing-task-description">
+                <h3><a href="#" class="link-regular">Подключить принтер</a></h3>
+                <p>Необходимо подключить старый матричный принтер, у него еще LPT порт…</p>
+            </div>
+            <div class="landing-task-info">
+                <div class="task-info-left">
+                    <p><a href="#" class="link-regular">Курьерские услуги</a></p>
+                    <p>25 минут назад</p>
+                </div>
+                <span>700 <b>₽</b></span>
+            </div>
+        </div>
+        <div class="landing-task">
+            <div class="landing-task-top task-cargo"></div>
+            <div class="landing-task-description">
+                <h3><a href="#" class="link-regular">Офисный переезд</a></h3>
+                <p>Требуется перевезти офисную мебель
+                    и технику из расчета 5 сотрудников</p>
+            </div>
+            <div class="landing-task-info">
+                <div class="task-info-left">
+                    <p><a href="#" class="link-regular">Грузоперевозки</a></p>
+                    <p>25 минут назад</p>
+                </div>
+                <span>1 800 <b>₽</b></span>
+            </div>
+        </div>
+        <div class="landing-task">
+            <div class="landing-task-top task-neo"></div>
+            <div class="landing-task-description">
+                <h3><a href="#" class="link-regular">Убраться в квартире</a></h3>
+                <p>Моей хате давно нужна генеральная уборка.
+                    В наличии есть только пылесос. </p>
+            </div>
+            <div class="landing-task-info">
+                <div class="task-info-left">
+                    <p><a href="#" class="link-regular">Уборка</a></p>
+                    <p>1 час назад</p>
+                </div>
+                <span>2000 <b>₽</b></span>
+            </div>
+        </div>
+        <div class="landing-task">
+            <div class="landing-task-top task-flat"></div>
+            <div class="landing-task-description">
+                <h3><a href="#" class="link-regular">Празднование ДР</a></h3>
+                <p>Моему другу нужно
+                    устроить день рождения,
+                    который он никогда не
+                    забудет</p>
+            </div>
+            <div class="landing-task-info">
+                <div class="task-info-left">
+                    <p><a href="#" class="link-regular">Мероприятия</a></p>
+                    <p>1 час назад</p>
+                </div>
+                <span>2000 <b>₽</b></span>
+            </div>
+        </div>
+    </div>
+    <div class="landing-bottom-container">
+        <button type="button" class="button red-button">смотреть все задания</button>
+    </div>
+</div>
+
