+++ frontend/views/layouts/main.php
@@ -33,7 +33,7 @@ AppAsset::register($this);
         <header class="page-header">
             <div class="main-container page-header__container">
                 <div class="page-header__logo">
-                    <a href="/">
+                    <a href="<?= Yii::$app->getHomeUrl()?>">
                         <svg class="page-header__logo-image" id="Layer_2" xmlns="http://www.w3.org/2000/svg"
                              viewBox="0 0 1634 646.35">
                             <title>taskforce_logo2-01</title>
@@ -143,7 +143,7 @@ AppAsset::register($this);
                                 <a href="#">Настройки</a>
                             </li>
                             <li>
-                                <a href="/user/logout">Выход</a>
+                                <a href="/users/logout">Выход</a>
                             </li>
                         </ul>
                     </div>
