+++ frontend/web/js/loginForm.js
@@ -1,26 +1,57 @@
-"use strict";
+'use strict';
+$(function() {
 
-let form = $("form#loginForm");
-form.submit(function(event) {
-    let data = form.serialize();
+    $('body').on("submit", "form#loginForm", function(event) {
+        let form = $(this);
+        console.log('sdfsd');
+        let data = form.serialize();
 
-    $('.help-block.help-block-error').text('').css( "margin-bottom", "0" );
-    $.ajax({
-        url: form.attr('action'),
-        type: 'POST',
-        data: data,
-        success: function (data) {
-            for (let prop in data) {
-                $('.field-'+prop + '+p').text(data[prop].join(', ')).css({
-                    "color": "red",
-                    "margin-bottom": "15px"
-                });
+        $('.help-block.help-block-error').text('').css('margin-bottom', '0');
+        $.ajax({
+            url: form.attr('action'),
+            type: 'POST',
+            data: data,
+            success: function(data) {
+                console.log(data);
+                for (let prop in data) {
+                    $('.field-' + prop + '+p').text(data[prop].join(', ')).css({
+                        'color': 'red',
+                        'margin-bottom': '15px'
+                    });
+                }
+            },
+            error: function(jqXHR, errMsg) {
+                alert(errMsg);
             }
-        },
-        error: function(jqXHR, errMsg) {
-            // alert(errMsg);
-        }
+        });
+
+        return false;
     });
 
-    return false;
-});
+ });
+// $(function() {
+// let form = $("form#loginForm");
+// form.submit(function() {
+//     let data = form.serialize();
+//
+//     $('.help-block.help-block-error').text('').css( "margin-bottom", "0" );
+//     $.ajax({
+//         url: form.attr('action'),
+//         type: 'POST',
+//         data: data,
+//         success: function (data) {
+//             for (let prop in data) {
+//                 $('.field-'+prop + '+p').text(data[prop].join(', ')).css({
+//                     "color": "red",
+//                     "margin-bottom": "15px"
+//                 });
+//             }
+//         },
+//         error: function(jqXHR, errMsg) {
+//             // alert(errMsg);
+//         }
+//     });
+//
+//     return false;
+// });
+// });
